requirements.txt
fastapi 
uvicorn 
motor 
langchain 
langchain-google-genai 
pydantic 
python-dotenv
langchain-core
#streamlit
requests


#rag
faiss-cpu
langchain-community
pypdf2
pypdf
python-docx
google-genai